try make RateFragment mvp
