+++ sprint5_Android/sprintDemo/app/src/main/java/com/example/user/text/rate/RateView.java
@@ -1,0 +1,12 @@
+package com.example.user.text.rate;
+
+import com.github.mikephil.charting.data.LineData;
+
+/**
+ * Created by ASUS on 2017/10/13.
+ */
+
+public interface RateView {
+
+   public void updateRate(LineData lineData);
+}
